### 2017-01-28 #94 @sckott
charlatan
`charlatan` makes fake data. It borrows data and modifies code from Python's [faker](https://github.com/joke2k/faker) library. `charlatan` adds more fake data types, in particular to target core audience of scientists.

locale support - done for a few data types; more work to be done on others. Most complete locale support in `JobProvider`/`ch_job` and `PersonProvider`/`name`.

```
Package: charlatan
Title: Make Fake Data
Description: Make fake data, supporting addresses, person names, dates,
    times, colors, coordinates, currencies, digital object identifiers
    (DOIs), jobs, phone numbers, DNA sequences, doubles and integers
    from distributions and within a range.
Version: 0.0.6.9560
Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
    email = "myrmecocystus@gmail.com"))
    R6 (>= 2.2.0),
    tibble (>= 1.2),
    whisker,
    Rmpfr
    covr,
Collate:
    'base-provider.R'
    'datetime-provider.R'
    'address-provider.R'
    ...
```

bottom of `Collate` cut off for brevity.

https://github.com/ropenscilabs/charlatan

Students in a classroom learning R, or e.g, in a software carpentry type bootcamp - since it's liteweight wrt dependencies, should be easy install everywhere. Also, if you're doing modeling/simulations, perhaps you could use this to generate some fake data to run models on.

not that i know of

Note: This package isn't quite complete in the sense that there's more types of data to add to the package, but the overall package API is set (but can be changed if needed), and tests are there (though more needed) and vignette exists - Anyway, submitting earlyish to get reviewer feedback
