### 2017-02-09 #94 @noamross
### Editor checks:

---

#### Editor comments

- I note that `ch_names()` can sometimes generate just last names, with blank spaces in front of them. Is this expected behavior? You may want to test for expected name patterns.
- Here is `goodpractice::gp()` output. Nothing major here, but a few small clean-ups:

── GP charlatan ────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 37% of code
    lines are covered by test cases.

    R/address-provider.R:89:NA
    R/address-provider.R:93:NA
    R/address-provider.R:97:NA
    R/address-provider.R:102:NA
    R/address-provider.R:103:NA
    ... and 267 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and
    developers are used it and it is easier to read your code for them if you use
    '<-'.

    R/color-provider.R:1:13
    R/credit_card-provider.R:14:17
    R/credit_card-provider.R:15:21
    R/credit_card-provider.R:16:19
    R/credit_card-provider.R:17:26
    ... and 103 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines shorter
    than 80 characters

    R/address-provider.R:122:1
    R/color-provider-uk_UA.R:6:1
    R/color-provider-uk_UA.R:7:1
    R/color-provider-uk_UA.R:9:1
    R/color-provider-uk_UA.R:11:1
    ... and 646 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...)
    expressions. They are error prone and result 1:0 if the expression on the right
    hand side is zero. Use seq_len() or seq_along() instead.

    R/missing-data-provider.R:26:19

  ✖ fix this R CMD check NOTE: Namespaces in Imports field not imported from: ‘R6’
    ‘Rmpfr’ ‘whisker’ All declared Imports should be used.
─────────────────────────────────────────────

`devtools::spell_check()` also found "inifinity" (infinity) and "dats" (data).
---

Reviewers:
Due date:
