### 2017-02-12 a6b8dd3
Added Hungarian name provider, base test case (to be amended) as well as some changes to the `Person` provider.
Hungarian names often contain abbreviated first, last or middle names - for instance, `D. Tóth Ferenc` has an abbreviated family name (in Hungarian, family names come before Christian names), used as `D.`, whereas others may have a middle abbreviation: e.g. `Kovács B. András`, where `B.` may stand for a middle name or for a second family name (usually maternal in case of children of divorced parents). To cater for this situation, I have added two class methods onto the Person provide
[…]
