+++ faker/providers/person/__init__.py
@@ -54,6 +54,18 @@ class Provider(BaseProvider):
         return cls.first_name()
 
     @classmethod
+    def first_name_male_abbreviated(cls):
+        if hasattr(cls, 'first_names_male'):
+            return cls.random_element(cls.first_names_male)[0] + "."
+        return cls.first_name()[0]
+
+    @classmethod
+    def first_name_female_abbreviated(cls):
+        if hasattr(cls, 'first_names_female'):
+            return cls.random_element(cls.first_names_female)[0] + "."
+        return cls.first_name()[0]
+
+    @classmethod
     def last_name_male(cls):
         if hasattr(cls, 'last_names_male'):
             return cls.random_element(cls.last_names_male)
