### 2017-02-12 266582c
Moved class methods for names from the base class into the specific provider class to align with practice.
