### 2017-02-17 d9c804c
Created Martial (ma_MA) providers.
Used the `@martianify` decorator to create Martian providers for
address, company, internet, job, lorem and person. Used deasciiify
directly to create a provider for color since I needed to keep the color
code un-martianified.
