+++ faker/providers/job/ma_MA/__init__.py
@@ -1,0 +1,7 @@
+from ..en_US import Provider as JobProvider
+from faker.utils.decorators import martianify
+
+
+@martianify('jobs')
+class Provider(JobProvider):
+    pass
