+++ faker/providers/lorem/ma_MA/__init__.py
@@ -1,0 +1,7 @@
+from ..la import Provider as LoremProvider
+from faker.utils.decorators import martianify
+
+
+@martianify('word_list')
+class Provider(LoremProvider):
+    pass
