+++ faker/providers/address/hu_HU/__init__.py
@@ -1,0 +1,144 @@
+# coding=utf-8
+
+from __future__ import unicode_literals
+from collections import OrderedDict
+
+from .. import Provider as AddressProvider
+
+
+class Provider(AddressProvider):
+    street_suffixes = OrderedDict(
+            (('utca', 0.75), ('út', 0.1), ('tér', 0.1), ('köz', 0.001), ('körút', 0.001), ('sétány', 0.001),))
+
+    street_name_formats = ('{{frequent_street_name}} {{street_suffix}}', '{{real_city_name}}i {{street_suffix}}',
+                           '{{city_part}}{{city_suffix}}i {{street_suffix}}',
+                           '{{city_prefix}}{{city_part}}i {{street_suffix}}')
+
+#   Currently deprecated.
+#   secondary_address_formats = ("#.em #.", "##. em. #.")
+
+    city_formats = ('{{city_prefix}}{{city_part}}{{city_suffix}}', '{{city_part}}{{city_suffix}}', '{{real_city_name}}')
+
+    # street_address_with_county_formats = (
+    #     '{{street_name}} {{building_number}}\n{{county_name}} megye\n{{postcode}} {{city}}',)
+
+    street_address_formats = ('{{street_name}} {{building_number}}',)
+
+    address_formats = ("{{street_address}}\n{{postcode}} {{city}}",)
+
+    frequent_street_names = (
+        'Ady Endre', 'Dózsa György', 'Petőfi', 'Petőfi Sándor', 'Arany János', 'Béke', 'Szabadság', 'Kossuth', 'József Attila')
+
+    # The 'real city name' generator includes a number of real cities of
+    # Hungary that no generator could feasibly dispense. Please note that the
+    # post code generator is, at this point, not capable of generating a
+    # fitting post code. In Hungary, post codes are determined by the county of
+    # the place (see the county generator), and for this reason, often there
+    # will be a discrepancy. A patch is in the works - until then, use
+    # Wikipedia to resolve postcode issues.
+    #
+    # This generator was created by collecting the 30 largest Hungarian places
+    # by population, based on the Hungarian Gazetteer generated with effect as
+    # of 01 January 2016 (http://www.ksh.hu/docs/hun/hnk/hnk_2016.pdf).
+
+    real_city_names = (
+        'Budapest', 'Debrecen', 'Szeged', 'Miskolc', 'Pécs', 'Győr', 'Nyíregyháza', 'Kecskemét', 'Székesfehérvár',
+        'Szombathely', 'Szolnok', 'Tatabánya', 'Érd', 'Kaposvár', 'Sopron', 'Veszprém', 'Békéscsaba', 'Zalaegerszeg',
+        'Eger', 'Nagykanizsa', 'Dunaújváros', 'Hódmezővásárhely', 'Dunakeszi', 'Szigetszentmiklós', 'Cegléd', 'Baja',
+        'Salgótarján', 'Ózd', 'Vác', 'Mosonmagyaróvár')
+
+    city_prefs = (
+        'kis', 'nagy', 'szent', 'duna', 'tisza', 'alsó', 'felső', 'belső', 'bakony', 'vác', 'mező', 'nyék', 'nyír',
+        'balaton', 'borsod', 'buda', 'hajdú', 'kun', 'moson', 'pilis', 'új', 'egyházas', 'dráva', 'magyar', 'mátra',
+        'somogy', 'lajos', 'bács', 'békés', 'puszta', 'orosz', 'rác', 'szerb', 'német', 'török')
+
[…]
