### 2017-02-18 0b91785
Hungarian SSN equivalent ('személyi szám') generator - even accepts birth dates and checks for new vs old checksum regimes!
