### 2017-02-18 bf00960
Documented Hungarian SSN generator.
