### 2017-02-19 31da27c
[ADD] Adding es_VE to language list
