+++ faker/providers/company/es_VE/__init__.py
@@ -1,0 +1,169 @@
+# -*- encoding: utf-8 -*-
+from __future__ import unicode_literals
+from .. import Provider as CompanyProvider
+
+
+class Provider(CompanyProvider):
+    formats = (
+        '{{last_name}} {{company_suffix}}',
+        '{{last_name}}-{{last_name}}',
+        '{{company_prefix}} {{last_name}}-{{last_name}}',
+        '{{company_prefix}} {{last_name}} y {{last_name}}',
+        '{{company_prefix}} {{last_name}}, {{last_name}} y {{last_name}}',
+        '{{last_name}}-{{last_name}} {{company_suffix}}',
+        '{{last_name}}, {{last_name}} y {{last_name}}',
+        '{{last_name}} y {{last_name}} {{company_suffix}}'
+    )
+
+    catch_phrase_words = (
+        (
+           "habilidad", "acceso", "adaptador", "algoritmo", "alianza",
+           "analista", "aplicación", "enfoque", "arquitectura",
+           "archivo", "inteligencia artificial", "array", "actitud",
+           "medición", "gestión presupuestaria", "capacidad", "desafío",
+           "circuito", "colaboración", "complejidad", "concepto",
+           "conglomeración", "contingencia", "núcleo", "fidelidad",
+           "base de datos", "data-warehouse", "definición", "emulación",
+           "codificar", "encriptar", "extranet", "firmware",
+           "flexibilidad", "focus group", "previsión", "base de trabajo",
+           "función", "funcionalidad", "Interfaz Gráfica", "groupware",
+           "Interfaz gráfico de usuario", "hardware", "Soporte", "jerarquía",
+           "conjunto", "implementación", "infraestructura", "iniciativa",
+           "instalación", "conjunto de instrucciones", "interfaz",
+           "intranet", "base del conocimiento", "red de area local",
+           "aprovechar", "matrices", "metodologías", "middleware",
+           "migración", "modelo", "moderador", "monitorizar",
+           "arquitectura abierta", "sistema abierto", "orquestar",
+           "paradigma", "paralelismo", "política", "portal",
+           "estructura de precios", "proceso de mejora",
+           "producto", "productividad", "proyecto", "proyección",
+           "protocolo", "línea segura", "software", "solución",
+           "estandardización", "estrategia", "estructura", "éxito",
+           "superestructura", "soporte", "sinergia", "mediante",
+           "marco de tiempo", "caja de herramientas", "utilización",
+           "website", "fuerza de trabajo"),
+        (
+            "24 horas", "24/7", "3ra generación", "4ta generación",
+            "5ta generación", "6ta generación", "analizada",
+            "asimétrica", "asíncrona", "monitorizada por red",
+            "bidireccional", "bifurcada", "generada por el cliente",
+            "cliente servidor", "coherente", "cohesiva", "compuesto",
+            "sensible al contexto", "basado en el contexto",
+            "basado en contenido", "dedicada",
+            "generado por la demanda", "didactica", "direccional",
+            "discreta", "dinámica", "potenciada", "acompasada",
+            "ejecutiva", "explícita", "tolerante a fallos",
+            "innovadora", "amplio ábanico", "global", "heurística",
+            "alto nivel", "holística", "homogénea", "híbrida",
+            "incremental", "intangible", "interactiva", "intermedia",
+            "local", "logística", "maximizada", "metódica",
+            "misión crítica", "móbil", "modular", "motivadora",
+            "multimedia", "multiestado", "multitarea", "nacional",
+            "basado en necesidades", "neutral", "nueva generación",
[…]
