+++ faker/providers/ssn/es_VE/__init__.py
@@ -1,0 +1,7 @@
+# coding=utf-8
+from __future__ import unicode_literals
+from .. import Provider as BaseProvider
+
+
+class Provider(BaseProvider):
+    ssn_formats = ("J-########-#",)
