+++ faker/providers/person/es_VE/__init__.py
@@ -1,0 +1,225 @@
+# coding=utf-8
+from __future__ import unicode_literals
+from .. import Provider as PersonProvider
+
+
+class Provider(PersonProvider):
+    formats_male = (
+        '{{first_name_male}} {{last_name}} {{last_name}}',
+        '{{first_name_male}} {{last_name}} {{last_name}}',
+        '{{first_name_male}} {{last_name}} {{last_name}}',
+        '{{first_name_male}} {{last_name}} {{last_name}}',
+        '{{first_name_male}} {{last_name}} {{last_name}}',
+        '{{first_name_male}} {{last_name}} {{last_name}}',
+        '{{first_name_male}} {{last_name}}',
+        '{{first_name_male}} {{prefix}} {{last_name}}',
+        '{{first_name_male}} {{last_name}}-{{last_name}}',
+        '{{first_name_male}} {{first_name_male}} {{last_name}} {{last_name}}',
+    )
+
+    formats_female = (
+        '{{first_name_female}} {{last_name}} {{last_name}}',
+        '{{first_name_female}} {{last_name}} {{last_name}}',
+        '{{first_name_female}} {{last_name}} {{last_name}}',
+        '{{first_name_female}} {{last_name}} {{last_name}}',
+        '{{first_name_female}} {{last_name}} {{last_name}}',
+        '{{first_name_female}} {{last_name}} {{last_name}}',
+        '{{first_name_female}} {{last_name}}',
+        '{{first_name_female}} {{prefix}} {{last_name}}',
+        '{{first_name_female}} {{last_name}}-{{last_name}}',
+        '{{first_name_female}} {{first_name_female}} {{last_name}} {{last_name}}',
+    )
+
+    formats = formats_male + formats_female
+
+    first_names_male = (
+        'Antonio', 'Jose', 'Manuel', 'Francisco', 'Juan', 'David', 'Jose Antonio', 'Jose Luis',
+        'Javier', 'Jesús', 'Francisco Javier', 'Daniel', 'Carlos', 'Miguel', 'Rafael', 'Jose Manuel',
+        'Alejandro', 'Pedro', 'Ángel', 'Miguel Ángel', 'Jose Maria', 'Fernando', 'Luís', 'Pablo',
+        'Sergio', 'Jorge', 'Alberto', 'Juan Carlos', 'Juan José', 'Diego', 'Alvaro', 'Adrián',
+        'Juan Antonio', 'Enrique', 'Raul', 'Ramon', 'Vicente', 'Ivan', 'Ruben', 'Andrés',
+        'Joaquin', 'Óscar', 'Juan Manuel', 'Santiago', 'Eduardo', 'Victor', 'Roberto', 'Jaime',
+        'Francisco Jose', 'Ignacio', 'Mario', 'Alfonso', 'Salvador', 'Ricardo', 'Jordi', 'Marcos',
+        'Emilio', 'Julian', 'Julio', 'Guillermo', 'Tomás', 'Gabriel', 'Agustín', 'Jose Miguel',
[…]
