+++ faker/providers/phone_number/es_VE/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import unicode_literals
+from .. import Provider as PhoneNumberProvider
+
+
+class Provider(PhoneNumberProvider):
+    formats = (
+        '+58 ### ### ## ##',
+        '+58 ### #######',
+        '+58(###)-#######',
+        '+58##########',
+    )
