+++ faker/providers/internet/es_VE/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+from __future__ import unicode_literals
+from .. import Provider as InternetProvider
+
+
+class Provider(InternetProvider):
+    tlds = (
+        'com.ve', 'com.ve', 'com.ve', 'com.ve', 'info.ve', 'net.ve', 'org.ve'
+    )
