### 2017-02-19 7947288
[FIX] Defined method zipcode and zipcode_plus4
