### 2017-02-20 933cc55
Fixed issues in the Hungarian address generator that caused the tests to fail.
