+++ faker/tests/hu_HU/__init__.py
@@ -12,3 +12,11 @@ class hu_HU_FactoryTestCase(unittest.TestCase):
     def setUp(self):
         self.factory = Factory.create('hu_HU')
 
+    def test_hungarian_postcodes_for_consistency(self):
+        # Hungarian postcodes may not begin with a zero. This test case tests for consistency with that rule.
+
+        from faker.providers.address.hu_HU import Provider
+
+        for i in range(100):
+            pcd = Provider.postcode()
+            assert pcd[2] > 0
