+++ faker/providers/address/hu_HU/__init__.py
@@ -131,18 +131,18 @@ class Provider(AddressProvider):
         return cls.random_element(cls.frequent_street_names)
 
     def city(self):
-        return super().city().capitalize()
+        return super(Provider, self).city().capitalize()
 
     @classmethod
     def postcode(cls):
-        return "H-{}{}{}{}".format(super(Provider).random_digit_not_null(), super(Provider).random_digit(),
-                                   super(Provider).random_digit(),
-                                   super(Provider).random_digit())
+        return "H-{}{}{}{}".format(super(Provider, cls).random_digit_not_null(), super(Provider, cls).random_digit(),
+                                   super(Provider, cls).random_digit(),
+                                   super(Provider, cls).random_digit())
 
     def street_name(self):
         return super(Provider, self).street_name().upper()
 
     @classmethod
     def building_number(cls):
-        numeric_part = super(Provider).random_int(1, 250)
+        numeric_part = super(Provider, cls).random_int(1, 250)
         return str(numeric_part) + "."
