### 2017-02-21 5046a47
remove __loader__ hack in tests
