### 2017-02-21 60f951d
move tests outside package src
