### 2017-02-21 4cdc66c
move tests outside package src
