+++ docs/index.rst
@@ -314,7 +314,7 @@ Installing dependencies:
 
 .. code:: bash
 
-    $ pip install -r faker/tests/requirements.txt
+    $ pip install -r tests/requirements.txt
 
 Run tests:
 
@@ -326,7 +326,7 @@ or
 
 .. code:: bash
 
-    $ python -m unittest -v faker.tests
+    $ python -m unittest -v tests
 
 Write documentation for providers:
 
