+++ circle.yml
@@ -1,3 +1,3 @@
 dependencies:
   pre:
-    - pip install -r faker/tests/requirements.txt
+    - pip install -r tests/requirements.txt
