+++ tests/__init__.py
@@ -204,7 +204,7 @@ class FactoryTestCase(unittest.TestCase):
         output = StringIO()
         print_doc(output=output)
         print_doc('address', output=output)
-        print_doc('faker.providers.person', output=output)
+        print_doc('faker.providers.person.it_IT', output=output)
         assert output.getvalue()
         with self.assertRaises(AttributeError):
             self.generator.get_formatter('barFormatter')
@@ -225,7 +225,7 @@ class FactoryTestCase(unittest.TestCase):
         orig_stdout = sys.stdout
         try:
             sys.stdout = StringIO()
-            command = Command(['faker', 'foo', '-i', 'tests.mymodule'])
+            command = Command(['faker', 'foo', '-i', 'tests.mymodule.en_US'])
             command.execute()
             assert sys.stdout.getvalue()
         finally:
