### 2017-02-21 9c0b73c
Revert "assume all providers are localized"
This reverts commit 95bbba9e00fa9ba33482a87767185049537653f7.

# Conflicts:
#	tests/__init__.py
