### 2017-02-24 1668867
Close #467. Include json files in distribution
