### 2017-02-24 180dc68
add hungarian to docs
