### 2017-02-24 c6251eb
Bump version: 0.7.7 → 0.7.8
