### 2017-02-24 6497049
Bump version: 0.7.7 → 0.7.8
