### 2017-02-26 e536ee2
Close #468. Exclude tests from distribution
