+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+`0.7.9 - 24-February-2017 <https://github.com/joke2k/faker/compare/v0.7.8...v0.7.9>`__
+--------------------------------------------------------------------------------------
+
+* Fix packaging isssue. Thanks @jorti.
+
 `0.7.8 - 24-February-2017 <https://github.com/joke2k/faker/compare/v0.7.7...v0.7.8>`__
 --------------------------------------------------------------------------------------
 
