### 2017-02-26 a54e68c
Bump version: 0.7.8 → 0.7.9
