### 2017-02-26 63a490b
Bump version: 0.7.8 → 0.7.9
