### 2017-03-07 #94 @geanders
## Package Review

- [x ] As the reviewer I confirm that there are no conflicts of interest for me to review this work (such as being a major contributor to the software).

#### Documentation

- [x ] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [ x] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [x ] **Function Documentation:** for all exported functions in R help
- [x ] **Examples** for all exported functions in R Help that run successfully locally
- [x ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and `URL`, `Maintainer` and `BugReports` fields in DESCRIPTION

#### Functionality

- [x ] **Installation:** Installation succeeds as documented.
- [ x] **Functionality:** Any functional claims of the software been confirmed.
- [ x] **Performance:** Any performance claims of the software been confirmed.
- [ x] **Automated tests:** Unit tests cover essential functions of the package
- [ x] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

#### Final approval (post-review)

Estimated hours spent reviewing: 4

---

### Review Comments

This package allows a user to create synthetic data of names, addresses, jobs, phone numbers, credit card numbers, taxonomies, etc., in several different languages. The package offers two interfaces, either through `fraudster` or through the `ch_*` family of functions. I can imagine this package being very useful for R package maintainers who are trying to develop R packages that wrap open data APIs, as it would allow the package maintainer to test package functions against data that might be input from the API.

There are cases where unreasonable values will be generated by the functions in this package (e.g., phone numbers with illegal area codes, weird names for people, etc.), which is to some extent unavoidable when generating synthetic data. Depending on the intended applications of this package, this may or may not be important. It would be useful to get an idea of more of the intended uses of this package to help assess if any of these unreasonable values (some of which I point out in comments below) would be likely to introduce problems when using this package.

The package name is great (and memorable). Coding style is excellent throughout. The vignette rendered without error locally and all examples also ran without error.

## Functions

- The use of a common prefix for all the `ch_` functions is very useful and will make it easier for users to use command completion to search for a function that want to use from the package but can't remember the name of.
- The `ch_*` functions are very clearly named.
- The ability to specify a locale when generating some types of data like names is particularly appealing. However, this will be much more helpful is there is a list provided somewhere (vignette maybe?) linking the possible choices for the `locale` code to use in this option to a location name. Also, I think that there is a typo in one of the options listed for local: I tried to google "hr_FR" (listed as an option for `ch_name` and `ch_job`) and couldn't come up with anything useful. Should this have been "fr_FR", which is already listed?
- The `ch_generate` function, which allows a user to create a dataframe with columns generated using other `ch_*` functions, is a very useful addition. However, the line `# ch_generate('job', 'color_name', 'stuff')` in the examples of the helpfile is confusing. Could this be removed?
- The package functions generally seem to have good checks of user inputs and helpful error messages if a user inputs something incorrect.
- Many of the `ch_*` functions used for columns allowed in `ch_generate` can be generated for different locales, but `ch_generate` itself does not allow a `locale` argument. Could it? (I would not mark this as a major concern, since a user could create a dataframe using several calls to separate `ch_*` functions.)
- Under the `es_ES` locale setting, it looks like many names get the same last name twice. Should this instead be sampling two different names from the last names vector and then putting them together? For example, the function currently gives values like this:

```
ch_name(n = 3, locale = "es_ES")
[1] "Nieves Nieves Diaz Diaz" "Sara Girón Girón"        "Clara Villena Villena"
```

- This may not be terribly important for the applications this package would be used for, but the method for names in "fr_FR" of adding random prefixes to some names is going to result in some funky last names in that language ("de la Dufour", "du Olivier", "Le Lemonnier"). Also, why do none of the French last names have accents in them, while the first names do? Will this matter in anticipated applications?
- For some of these locales (for example, "fr_FR"), many job titles would be different depending on whether the worker is male or female. In the applications you envision for this package, will it be a problem that all the jobs generated with `ch_job` are for male workers?
- Could the phone number generator include locale as an option?
- Are there any advantages to using the functions enabled by `numerics-provider.R` (e.g. `ch_norm`, `ch_unif`, `ch_lnorm`, `ch_beta`) beyond using the base R analogs (`rnorm`, etc.)? If there are important advantages, it would be helpful if they were explained in the helpfile for these packages. If not, are they necessary to include in this package?
- It may be useful to allow the user to set ranges for `ch_date_time`, similar to how a bounding box can be set for `ch_position`.
- There are a number of files in the `R` subdirectory of the package for creating company names, but there doesn't seem to be a `ch_company` function that uses that. Is the company provider meant to be used anywhere by a package user? If not, could this code be removed from the package, or moved to a development branch?

## Vignette

- The formatting of the vignette, with a linked table of contents to the left, is very attractive. However, the text of the vignette is clearly still unfinished. Given that the intended use of the package is pretty straightforward, the vignette doesn't need to be very detailed, but it would be helpful if it included a short paragraph at the beginning introducing the package, for example, as well as a paragraph explaining the differences between using `fraudster` versus the `ch_*` family of functions (even just adding the information currently in the helpfile for the package would be useful).
- It would be very useful to have more details, either in the vignette or in the helpfiles, about how the sythetic data is created and how much variety is available for each category. In particular, how were the values picked that were used in generating vectors for names, jobs, etc., for each locale (e.g., in `person-provider-en_US.R`)? Are these representative of common names in each locale? The details provided in the helpfile for `TaxonomyProvider` are very useful and the perfect level of detail-- something similar for other providers, and for the `ch_*` functions that use them, would be great.
- For `ch_job`, it looks like the "en_US" locale values are more consistent with a "en_GB" locale. British spellings are used in these job titles ("maths", "centre"). Does this matter?
- "available" is misspelled in the error message "not in set of avaiable locales".
- For the "en_US" last names, at least, it doesn't look like there are any two-word last names (e.g., "von Trapp"). If this package is used to create test data, I think some cases of two-word last names could be useful.
- Again, if this package is meant, at least as one use, to create "messy" data to use for function tests, it could be useful to introduce some variation in honorifics and degree listings (e.g., both "Mr" and "Mr.", both "PhD" and "Ph.D.") that might show up in real data.

## Helpfiles

- Check for typos throughout.
- Based on the helpfiles for the `ch_*` functions, the `n` parameter can be "to infinity". I found this wording a bit awkard, as the the functions (expectedly) return an error if run with `n` set to `Inf`: `Error in integer(n) : vector size cannot be infinite`. Perhaps reword this in the helpfile to something like "any positive integer"?
- In the helpfile for `ch_generate`, it would be useful to include a note on the columns that are included by default (name, job, and phone number). Also, there's currently text that suggests to "See Details", but no Details section.
