### 2017-03-09 9141b3d
Implement ISBN-13
