### 2017-03-09 b987e87
Implement ISBN-10
