+++ faker/providers/isbn/isbn.py
@@ -38,3 +38,25 @@ class ISBN13(ISBN):
     def format(self, separator=''):
         return separator.join([self.ean, self.group, self.registrant,
                                self.publication, self.check_digit])
+
+
+class ISBN10(ISBN):
+
+    def __init__(self, *args, **kwargs):
+        super(ISBN10, self).__init__(*args, **kwargs)
+        self.check_digit = self._check_digit()
+
+    def _check_digit(self):
+        """ Calculate the check digit for ISBN-10.
+        See https://en.wikipedia.org/wiki/International_Standard_Book_Number
+        for calculation.
+        """
+        weights = range(1, 10)
+        body = ''.join([self.group, self.registrant, self.publication])
+        remainder = sum(int(b) * w for b, w in zip(body, weights)) % 11
+        check_digit = 'X' if remainder == 10 else str(remainder)
+        return str(check_digit)
+
+    def format(self, separator=''):
+        return separator.join([self.group, self.registrant, self.publication,
+                               self.check_digit])
