### 2017-03-09 #94 @sckott
response to @tjmahr - thx much for your review!

---------

#### BaseProvider

> This package makes extensive use of the `sample()` function. This function has a flawed default behavior: It expands integers into ranges. Thus, `sample(10, 1)` is the same as `sample(1:10, 1)`. This package's `BaseProvider$random_element()` method inherits this flaw. Perhaps, `x[sample(seq_along(x), 1)]`, which samples from a sequence of item positions, would be a better implementation for this method.

see https://github.com/ropenscilabs/charlatan/issues/12 - looking into it

> The method `random_int()` has the signature ... It can never generate the maximum value, and this fact isn't noted anywhere.

see https://github.com/ropenscilabs/charlatan/issues/13 - good catch

> I'm curious why the digit-or-blank generators don't just sample from a larger set. The code for `random_digit_not_null_or_empty()` means that half of the generated items should be `""`. Why not sample with this instead? ...

see https://github.com/ropenscilabs/charlatan/issues/14 - good catch, will fix

#### NumericsProvider

>  The `integer()` method doesn't defer to R, but it instead uses `sample()`. By default `sample()` does not sample with replacement, so `NumericsProvider$new()$integer()` and `ch_integer()` can behave in unexpected ways.

This is b/c i was following what `faker` lib was doing - it's based on picking a random integer from a range given bay a user - not from some distribution

> It is also odd that the `random_int()` and `integer()` use different techniques for generating random integers.

see above comment

#### PersonProvider

> As the example above shows, the names can vary in format. (Sometimes there is a blank first name --- this might be a bug.)

see https://github.com/ropenscilabs/charlatan/issues/15  looking into it

>  The locale-specific names are stored as variables in the package's namespace, and they are stored in R scripts. Thus, there are .R files that contain vectors with thousands of names. I wonder if using a `data/` folder would be a cleaner way to separate code and package data.

I've thought about storing as data - as binary I don't like as the data may need to be updated, changed - so possibly `data-raw` instead - i can look into it

> The documentation for `ch_name()` under-reports the supported locales. For example, Spanish is supported but this Spanish support is not documented.

right, a symptom of submitting this for review early in development :) will fix

> There is a bug in how double last names (as in Spanish) are generated.

see https://github.com/ropenscilabs/charlatan/issues/16 thanks, fixing

#### ColorProvider

> Its hex-color generator users the `random_int()` method to choose a number between 1 and 16777215. The above noted behavior means it will never generate \#ffffff, and the lower bound of 1 prevents \#000000 from being generated too.

will fix see https://github.com/ropenscilabs/charlatan/issues/18

> Some locale sensitivity for color names, but not safe ones

thanks, will look into it, see https://github.com/ropenscilabs/charlatan/issues/17

> It pads zeros onto strings less than 6 characters to create the familiar six- digit format. But padding is done on the right side, so it will never generate "\#0000ff". It also counts character length after appending the the pound sign, so the above example shows an incorrect hex color. ~~I think using `sprintf("#%06x", ...)` would fix these problems.~~ Actually, I discovered `grDevices::rgb()` while reading about the `colors()` for a later comment. That function should power the implementation.  A similar strategy underlies `safe_hex_color()`, so the method has the same flaws. Plus, something mysterious happens sometimes: ...

thanks, for all those see https://github.com/ropenscilabs/charlatan/issues/19

#### CoordinateProvider

> It can generate coordinates within a boundary box. The box is not checked, so users can get invalid coordinates as a result.

see https://github.com/ropenscilabs/charlatan/issues/20

#### CreditCardProvider

> ... there is some commented out Python/R code in place.

Right, just not done yet

#### DateTimeProvider

> This class powers `ch_timezone()`, `ch_unix_time()`, and `ch_date_time()`. It is not documented that the date-times are integers converted to POSIX times, so they are all after 1970.

will fix that, see https://github.com/ropenscilabs/charlatan/issues/21

> The file also features some commented out Python code.

Right, just not done yet

> There is a typo in `century()` method, so it always errors.

will fix, see https://github.com/ropenscilabs/charlatan/issues/22

#### TaxonomyProvider

> The basis for the random genus/species names is well detailed in the hidden `?TaxonomyProvider` page but not included in the user-facing `?taxonomy` page. Now that Roxygen2 supports the new `@inherit` and `@inheritSection fun title` directives, I suggest that the user-facing page include this documentation.

will do, see https://github.com/ropenscilabs/charlatan/issues/23

#### Do-one-thing generators

##### DOIProvider

> This class does not use its inherited `random_int()` method to generate the integers, but it should.

good idea, see https://github.com/ropenscilabs/charlatan/issues/24

##### SequenceProvider

> I feel the user-facing function should be named `ch_gene_sequence()`, not `ch_sequence()`

probably, see https://github.com/ropenscilabs/charlatan/issues/25

#### FraudsterClient

> Because locales are not uniformly supported, this can lead to errors.

right, again, a symptom of submitting early in dev for review, will try to fail better on that https://github.com/ropenscilabs/charlatan/issues/26

> Also, the `name()` method ignores the locale even though it can support different locales.

good catch, see https://github.com/ropenscilabs/charlatan/issues/26

#### Unfinished providers

> I didn't closely review the code for these classes, as they appear to be unfinished and are not user-facing like the other ones. I am documenting them here for completeness.

right, again symptom of submitting for review earlyish

> `company_provider` ... has a different R6 design and naming scheme than other classes

will fix that to make sure same format as others, see https://github.com/ropenscilabs/charlatan/issues/27

> `MissingProvider` provides a method to inject NA values into a vector. There is no user-facing function (something like `ch_missing()`) for this class).

will add that, see https://github.com/ropenscilabs/charlatan/issues/28

> it's odd that the *n* missing values it generates for a vector overwrite the first *n* values. I feel like it should randomly determine *n*, `sample()` *n* position indices, and make the elements in those positions NA so that the NAs are scattered throughout the vector.

see https://github.com/ropenscilabs/charlatan/issues/29

#### `ch_generate()`

> The standard error for a bad field name is ... A better error might say, `see ?ch_generate` or print out the `all_choices` vector.

will point user to `?ch_generate` - the list of options is short right now, but i think it will get quite long - in which case cumbersome to print all out, but maybe I will anyway  https://github.com/ropenscilabs/charlatan/issues/30

> `ch_generate("hex_color")` does not work, although the documentations claims to support it.  The documentation for the `n` argument says "any (integer) number of things to get, 1 to inifinity". I would recommend just saying any non-negative integer. This applies to every documentation page as well. The documentation for `...` should also say that `name`, `job`, and `phone_number` columns are created by default if nothing is specified.

will fix, https://github.com/ropenscilabs/charlatan/issues/31

#### Other comments

> `parse_eval()` should move to zzz.R, where the rest of the utility functions live.

done

> Personally, I would avoid using this function by structuring all the data into nested lists so that items can be retrieved with `data[[locale_name]]` instead of using `parse_eval()` to create-then-evaluate a variable name.

why?

> The `random_digit_not_null()` method in the BaseGenerator is better named `random_digit_not_zero()`.

good idea, done

> The unit test for colors and safe colors should check that the generated colors meets the color and safe-color formats.

thanks, done

> There are only unit tests for the colors, coordinates, credit cards, currency and jobs. I would suggest adding tests for the functions I've described bugs in.

definitely! meaning to fill out tests more, will work on that, https://github.com/ropenscilabs/charlatan/issues/2
