### 2017-03-09 #94 @sckott
response to @geanders  - thx much for your review!

-----------

> There are cases where unreasonable values will be generated by the functions in this package (e.g., phone numbers with illegal area codes, weird names for people, etc.), which is to some extent unavoidable when generating synthetic data.

@geanders  What do you mean by illegal?  e.g., with area codes, do you mean it generates an area code that doesn't exist?  the rules in the pkg for PhoneNumberProvider are only for the formats the string takes, not what combinations are in the real set.  I don't think our goal here is to create real phone numbers or real anything else. I think that would in fact not be a good idea. However, cases where "illegal" values are created with respect to not following a specific format then we'd want to fix that.

> Depending on the intended applications of this package, this may or may not be important. It would be useful to get an idea of more of the intended uses of this package to help assess if any of these unreasonable values (some of which I point out in comments below) would be likely to introduce problems when using this package.

I don't have any specific applications in mind. I think the pkg could be used in A LOT of different use cases: generating datasets in another pkg, datasets for teaching, pkg test suites, generating datasets with a certain structure for a very specific use case, etc. I'll add more to the docs about what possible use cases are https://github.com/ropenscilabs/charlatan/issues/32

## Functions

> The ability to specify a locale when generating some types of data like names is particularly appealing. However, this will be much more helpful is there is a list provided somewhere (vignette maybe?) linking the possible choices for the `locale` code to use in this option to a location name.

>Also, I think that there is a typo in one of the options listed for local: I tried to google "hr_FR" (listed as an option for `ch_name` and `ch_job`) and couldn't come up with anything useful. Should this have been "fr_FR", which is already listed?

sorry, that should be `hr_HR` (croation, http://www.localeplanet.com/icu/hr-HR/index.html), fixed in both

> The `ch_generate` function, which allows a user to create a dataframe with columns generated using other `ch_*` functions, is a very useful addition. However, the line `# ch_generate('job', 'color_name', 'stuff')` in the examples of the helpfile is confusing. Could this be removed?

sure, removed (commented out eg, to show what happens when there's an invalid variable input

> Many of the `ch_*` functions used for columns allowed in `ch_generate` can be generated for different locales, but `ch_generate` itself does not allow a `locale` argument. Could it? (I would not mark this as a major concern, since a user could create a dataframe using several calls to separate `ch_*` functions.)

sounds good to add it in there, see https://github.com/ropenscilabs/charlatan/issues/33

> Under the `es_ES` locale setting, it looks like many names get the same last name twice. Should this instead be sampling two different names from the last names vector and then putting them together?

Right, will fix, see https://github.com/ropenscilabs/charlatan/issues/34

> This may not be terribly important for the applications this package would be used for, but the method for names in "fr_FR" of adding random prefixes to some names is going to result in some funky last names in that language ("de la Dufour", "du Olivier", "Le Lemonnier"). Also, why do none of the French last names have accents in them, while the first names do? Will this matter in anticipated applications?

I'll have a look at this, but i'm only following what the Python library did, see https://github.com/joke2k/faker/blob/master/faker/providers/person/fr_FR/__init__.py I'll see why they don't have an accents in last names - see https://github.com/ropenscilabs/charlatan/issues/35

> For some of these locales (for example, "fr_FR"), many job titles would be different depending on whether the worker is male or female. In the applications you envision for this package, will it be a problem that all the jobs generated with `ch_job` are for male workers?

@geanders  What are examples of different titles for female vs. male?

> Could the phone number generator include locale as an option?

yep, just haven't gotten to it yet, as you can see at https://github.com/joke2k/faker/tree/master/faker/providers there's a lot in there I haven't gotten to yet in this pkg, see https://github.com/ropenscilabs/charlatan/issues/36

> Are there any advantages to using the functions enabled by `numerics-provider.R` (e.g. `ch_norm`, `ch_unif`, `ch_lnorm`, `ch_beta`) beyond using the base R analogs (`rnorm`, etc.)? If there are important advantages, it would be helpful if they were explained in the helpfile for these packages. If not, are they necessary to include in this package?

these were started from this discussion https://github.com/ropenscilabs/charlatan/issues/11 - yes, they aren't adding much value, but i think there's more to do with them

> It may be useful to allow the user to set ranges for `ch_date_time`, similar to how a bounding box can be set for `ch_position`.

right, datetime provider is very incomplete - there's many methods not added yet - including date times between a range of dates, see  https://github.com/joke2k/faker/blob/master/faker/providers/date_time/__init__.py#L362 - so it will be added in time

> There are a number of files in the `R` subdirectory of the package for creating company names, but there doesn't seem to be a `ch_company` function that uses that. Is the company provider meant to be used anywhere by a package user? If not, could this code be removed from the package, or moved to a development branch?

xxxx

## Vignette

- The formatting of the vignette, with a linked table of contents to the left, is very attractive. However, the text of the vignette is clearly still unfinished. Given that the intended use of the package is pretty straightforward, the vignette doesn't need to be very detailed, but it would be helpful if it included a short paragraph at the beginning introducing the package, for example, as well as a paragraph explaining the differences between using `fraudster` versus the `ch_*` family of functions (even just adding the information currently in the helpfile for the package would be useful).
- It would be very useful to have more details, either in the vignette or in the helpfiles, about how the sythetic data is created and how much variety is available for each category. In particular, how were the values picked that were used in generating vectors for names, jobs, etc., for each locale (e.g., in `person-provider-en_US.R`)? Are these representative of common names in each locale? The details provided in the helpfile for `TaxonomyProvider` are very useful and the perfect level of detail-- something similar for other providers, and for the `ch_*` functions that use them, would be great.
- For `ch_job`, it looks like the "en_US" locale values are more consistent with a "en_GB" locale. British spellings are used in these job titles ("maths", "centre"). Does this matter?
- "available" is misspelled in the error message "not in set of avaiable locales".
- For the "en_US" last names, at least, it doesn't look like there are any two-word last names (e.g., "von Trapp"). If this package is used to create test data, I think some cases of two-word last names could be useful.
- Again, if this package is meant, at least as one use, to create "messy" data to use for function tests, it could be useful to introduce some variation in honorifics and degree listings (e.g., both "Mr" and "Mr.", both "PhD" and "Ph.D.") that might show up in real data.

## Helpfiles

- Check for typos throughout.
- Based on the helpfiles for the `ch_*` functions, the `n` parameter can be "to infinity". I found this wording a bit awkard, as the the functions (expectedly) return an error if run with `n` set to `Inf`: `Error in integer(n) : vector size cannot be infinite`. Perhaps reword this in the helpfile to something like "any positive integer"?
- In the helpfile for `ch_generate`, it would be useful to include a note on the columns that are included by default (name, job, and phone number). Also, there's currently text that suggests to "See Details", but no Details section.
