### 2017-03-10 feede7d
Add french colors
