+++ faker/providers/color/fr_FR/__init__.py
@@ -1,0 +1,156 @@
+# coding=utf-8
+from __future__ import unicode_literals
+from collections import OrderedDict
+
+from .. import Provider as ColorProvider
+
+localized = True
+
+
+class Provider(ColorProvider):
+    all_colors = OrderedDict((
+        ("Noir", "#000000"),
+        ("Gris mat", "#696969"),
+        ("Gris", "#808080"),
+        ("Gris foncé (Acier)", "#A9A9A9"),
+        ("Gris argent", "#C0C0C0"),
+        ("Gris clair", "#D3D3D3"),
+        ("Gris gainsboro (Etain)", "#DCDCDC"),
+        ("Blanc fumée", "#F5F5F5"),
+        ("Blanc spectral", "#F8F8FF"),
+        ("Blanc", "#FFFFFF"),
+        ("Ivoire", "#FFFFF0"),
+        ("Blanc floral", "#FFFAF0"),
+        ("Blanc coquillage", "#FFF5EE"),
+        ("Blanc lavande", "#FFF0F5"),
+        ("Blanc dentelle", "#FDF5E6"),
+        ("Blanc Lin", "#FAF0E6"),
+        ("Rose brumeux", "#FFE4E1"),
+        ("Rose", "#FFC0CB"),
+        ("Rose clair", "#FFB6C1"),
+        ("Rose Passion", "#FF69B4"),
+        ("Rose profond", "#FF1493"),
+        ("Violet pâle", "#DB7093"),
+        ("Fushia (Magenta)", "#FF00FF"),
+        ("Violet moyen", "#C71585"),
+        ("Violet chardon", "#D8BFD8"),
+        ("Prune", "#DDA0DD"),
+        ("Violet", "#EE82EE"),
+        ("Violet orchidée", "#DA70D6"),
+        ("Violet orchidée moyen", "#BA55D3"),
+        ("Violet orchidée foncé", "#9932CC"),
+        (u"Violet foncé", "#9400D3"),
+        ("Bleu violet", "#8A2BE2"),
+        ("Indigo", "#4B0082"),
+        ("Bleu ardoise moyen", "#7B68EE"),
+        ("Bleu ardoise", "#6A5ACD"),
+        ("Bleu ardoise foncé", "#483D8B"),
+        ("Pourpre moyen", "#9370DB"),
+        ("Magenta foncé", "#8B008B"),
+        ("Pourpre", "#800080"),
+        ("Brun rosé", "#BC8F8F"),
+        ("Corail clair", "#F08080"),
+        ("Corail", "#FF7F50"),
+        ("Tomate", "#FF6347"),
+        ("Orangé", "#FF4500"),
+        ("Rouge", "#FF0000"),
+        ("Rouge cramoisi", "#DC143C"),
+        ("Saumon clair", "#FFA07A"),
+        ("Saumon Foncé", "#E9967A"),
+        ("Saumon", "#FA8072"),
+        ("Rouge Indien", "#CD5C5C"),
+        ("Rouge brique", "#B22222"),
+        ("Brun", "#A52A2A"),
+        ("Rouge foncé", "#8B0000"),
+        ("Bordeaux", "#800000"),
+        ("Beige", "#F5F5DC"),
+        ("Beige antique", "#FAEBD7"),
+        ("Beige papaye", "#FFEFD5"),
+        ("Amande", "#FFEBCD"),
+        ("Bisque", "#ffe4c4"),
+        ("Beige pêche", "#FFDAB9"),
+        ("Beige mocassin", "#FFE4B5"),
+        ("Jaune blanc navaro", "#FFDEAD"),
+        ("Jaune blé", "#F5DEB3"),
[…]
