### 2017-03-10 #94 @tjmahr
> > NumericsProvider
> >
> > The integer() method doesn't defer to R, but it instead uses sample(). By default sample() does not sample with replacement, so NumericsProvider$new()$integer() and ch_integer() can behave in unexpected ways.
>
> This is b/c i was following what faker lib was doing - it's based on picking a random integer from a range given bay a user - not from some distribution

Am I allowed to elaborate comments in a review? Anyway, this implementation of `ch_integer()` behaves unlike `ch_double()` or `ch_unif()` because there's a limit to the number of integers that can be generated when sampling without replacement.

```r
charlatan::ch_integer(100, min = 1, max = 10)
#> Error in sample.int(length(x), size, replace, prob) :
#>   cannot take a sample larger than the population when 'replace = FALSE'
```

This makes it behave in unexpected ways compared to its sibling functions. In this way, it's not really a random integer generator, although it's lumped in with those other number generators, but a function that shuffles a pile of numbers and draws from that shuffled pile.
