### 2017-03-11 ac5f7f6
Remove unused localized attribute
