### 2017-03-11 bb47e8b
Revert "Remove unused localized attribute"
This reverts commit ac5f7f64bee2b1b601fcd855bfaedbdd988b2f53.
