+++ faker/providers/color/fr_FR/__init__.py
@@ -4,6 +4,8 @@ from collections import OrderedDict
 
 from .. import Provider as ColorProvider
 
+localized = True
+
 
 class Provider(ColorProvider):
     all_colors = OrderedDict((
