### 2017-03-12 1be5cd2
Document supported Python versions
Documented all versions that are tested by Travis.
