+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.9
+current_version = 0.7.10
 files = setup.py faker/__init__.py docs/conf.py
 commit = True
 tag = True
