+++ faker/providers/person/th_TH/__init__.py
@@ -1,4 +1,4 @@
-# coding = 'utf8'
+# -*- coding: utf-8 -*-
 
 from __future__ import unicode_literals
 from .. import Provider as PersonProvider
