### 2017-03-14 330a2ae
Revert "add provider en_TH, th_TH"
