### 2017-03-14 4361f44
Reorganize hr_HR SSN tests
