### 2017-03-15 fe167bf
Fix incorrect ne_NP locale tests
This test incorrectly assumes a call to name() will
yield only a first/last name, which isn't always true for this
locale. I suspect it hasn't been uncovered yet because the
tests are seeded the same at the beginning of every run. It only
becomes a problem when you start moving tests around. This change
addresses the incorrect assertions as well as makes the file PEP8
compliant.
