+++ faker/providers/date_time/fr_FR/__init__.py
@@ -16,7 +16,8 @@ class Provider(DateTimeProvider):
             "3": "Mercredi",
             "4": "Jeudi",
             "5": "Vendredi",
-            "6": "Samedi",}
+            "6": "Samedi",
+        }
         return DAY_NAMES[day]
 
     @classmethod
@@ -34,6 +35,7 @@ class Provider(DateTimeProvider):
             "09": "Septembre",
             "10": "Octobre",
             "11": "Novembre",
-            "12": "Décembre",}
+            "12": "Décembre",
+        }
         return MONTH_NAMES[month]
 
