### 2017-03-15 5fd0a33
Reorganize hu_HU postcode test
