### 2017-03-15 910d273
Rename tests/postcodes to tests/addresses
This keeps test modules names in line with the actual module names.
