### 2017-03-16 2991506
Reorganize ja_JP address test
