+++ tests/providers/address.py
@@ -2,11 +2,15 @@
 
 from __future__ import unicode_literals
 
+import re
 import unittest
 
-from faker import Factory
 from ukpostcodeparser.parser import parse_uk_postcode
+
+from faker import Factory
 from faker.providers.address.hu_HU import Provider as HuProvider
+from faker.providers.address.ja_JP import Provider as JaProvider
+from .. import string_types
 
 
 class TestEnGB(unittest.TestCase):
@@ -29,3 +33,50 @@ class TestHuHU(unittest.TestCase):
         for i in range(100):
             pcd = HuProvider.postcode()
             assert pcd[2] > "0"
+
+
+class TestJaJP(unittest.TestCase):
+    """ Tests addresses in the ja_JP locale """
+
+    def setUp(self):
+        self.factory = Factory.create('ja')
+
+    def test_address(self):
+        country = self.factory.country()
+        assert isinstance(country, string_types)
+        assert country in JaProvider.countries
+
+        prefecture = self.factory.prefecture()
+        assert isinstance(prefecture, string_types)
+        assert prefecture in JaProvider.prefectures
+
+        city = self.factory.city()
+        assert isinstance(city, string_types)
+        assert city in JaProvider.cities
+
+        town = self.factory.town()
+        assert isinstance(town, string_types)
+        assert town in JaProvider.towns
+
+        chome = self.factory.chome()
+        assert isinstance(chome, string_types)
+        assert re.match("\d{1,2}丁目", chome)
+
+        ban = self.factory.ban()
+        assert isinstance(ban, string_types)
+        assert re.match("\d{1,2}番", ban)
+
+        gou = self.factory.gou()
+        assert isinstance(gou, string_types)
+        assert re.match("\d{1,2}号", gou)
+
+        building_name = self.factory.building_name()
+        assert isinstance(building_name, string_types)
+        assert building_name in JaProvider.building_names
+
+        zipcode = self.factory.zipcode()
+        assert isinstance(zipcode, string_types)
+        assert re.match("\d{3}-\d{4}", zipcode)
+
+        address = self.factory.address()
+        assert isinstance(address, string_types)
