### 2017-03-16 5b74723
Reorganize ja_JP company test
