### 2017-03-16 7a06509
Reorganize ja_JP internet test
