### 2017-03-16 1e3bafa
Reorganize ja_JP person test
