### 2017-03-16 b554a97
Reorganzie ja_JP phone_number test
