### 2017-03-16 71a2c80
Reorganize ne_NP address test
