### 2017-03-16 1b5f0b0
Reorganize ne_NP person test
