### 2017-03-16 6d37756
Reorganize no_NO address test
