### 2017-03-16 2fd2723
Reorganize pr_BR ssn tests
