+++ tests/providers/ssn.py
@@ -6,7 +6,8 @@ import unittest
 import re
 
 from faker import Factory
-from faker.providers.ssn.hr_HR import Provider, checksum
+from faker.providers.ssn.hr_HR import Provider as HrProvider, checksum as hr_checksum
+from faker.providers.ssn.pt_BR import Provider as PtProvider, checksum as pt_checksum
 
 
 class TestHrHR(unittest.TestCase):
@@ -16,13 +17,32 @@ class TestHrHR(unittest.TestCase):
         self.factory = Factory.create('hr_HR')
 
     def test_ssn_checksum(self):
-        self.assertEqual(checksum([0, 0, 2, 2, 8, 2, 6, 9, 2, 8]), 9)
-        self.assertEqual(checksum([5, 8, 9, 3, 6, 9, 5, 1, 2, 5]), 1)
-        self.assertEqual(checksum([5, 7, 8, 0, 2, 0, 3, 4, 2, 3]), 7)
-        self.assertEqual(checksum([4, 3, 3, 3, 1, 4, 6, 7, 6, 2]), 2)
-        self.assertEqual(checksum([0, 5, 9, 3, 7, 7, 5, 9, 1, 8]), 7)
-        self.assertEqual(checksum([7, 1, 1, 4, 9, 9, 1, 2, 4, 1]), 6)
+        self.assertEqual(hr_checksum([0, 0, 2, 2, 8, 2, 6, 9, 2, 8]), 9)
+        self.assertEqual(hr_checksum([5, 8, 9, 3, 6, 9, 5, 1, 2, 5]), 1)
+        self.assertEqual(hr_checksum([5, 7, 8, 0, 2, 0, 3, 4, 2, 3]), 7)
+        self.assertEqual(hr_checksum([4, 3, 3, 3, 1, 4, 6, 7, 6, 2]), 2)
+        self.assertEqual(hr_checksum([0, 5, 9, 3, 7, 7, 5, 9, 1, 8]), 7)
+        self.assertEqual(hr_checksum([7, 1, 1, 4, 9, 9, 1, 2, 4, 1]), 6)
 
     def test_ssn(self):
         for i in range(100):
-            self.assertTrue(re.search(r'^\d{11}$', Provider.ssn()))
+            self.assertTrue(re.search(r'^\d{11}$', HrProvider.ssn()))
+
+
+class TestPtBR(unittest.TestCase):
+
+    def setUp(self):
+        self.factory = Factory.create('pt_BR')
+
+    def test_pt_BR_ssn_checksum(self):
+        self.assertEqual(pt_checksum([8, 8, 2, 8, 2, 1, 6, 5, 2]), 2)
+        self.assertEqual(pt_checksum([8, 8, 2, 8, 2, 1, 6, 5, 2, 2]), 1)
+
+    def test_pt_BR_ssn(self):
+        for _ in range(100):
+            self.assertTrue(re.search(r'^\d{11}$', PtProvider.ssn()))
+
+    def test_pt_BR_cpf(self):
+        for _ in range(100):
+            self.assertTrue(re.search(r'\d{3}\.\d{3}\.\d{3}\-\d{2}', PtProvider.cpf()))
+
