### 2017-03-16 ea896b3
Reorganize pt_BR company tests
