### 2017-03-16 ae7c536
Reorganize zh_CN internet test
