### 2017-03-16 5bd427b
Move utils tests to own package
