### 2017-03-16 873436f
Move generator tests to own module
