+++ tests/__init__.py
@@ -12,11 +12,6 @@ import sys
 import logging
 
 try:
-    from mock import patch
-except ImportError:  # pragma: no cover
-    from unittest.mock import patch
-
-try:
     from StringIO import StringIO
 except ImportError:  # pragma: no cover
     from io import StringIO
@@ -576,23 +571,6 @@ class FactoryTestCase(unittest.TestCase):
 
         number = provider.random_number(10, True)
         self.assertEqual(len(str(number)), 10)
-
-
-class GeneratorTestCase(unittest.TestCase):
-
-    def setUp(self):
-        self.generator = Generator()
-
-    @patch('random.getstate')
-    def test_get_random(self, mock_system_random):
-        random_instance = self.generator.random
-        random_instance.getstate()
-        self.assertFalse(mock_system_random.called)
-
-    @patch('random.seed')
-    def test_random_seed_doesnt_seed_system_random(self, mock_system_random):
-        self.generator.seed(0)
-        self.assertFalse(mock_system_random.called)
 
 
 if __name__ == '__main__':
