### 2017-03-16 e9472fb
Move factory tests to own module
