+++ faker/providers/person/th_TH/__init__.py
@@ -10,7 +10,11 @@ class Provider(PersonProvider):
         '{{first_name}} {{last_name}}',
         '{{first_name}} {{last_name}}',
         '{{first_name}} {{last_name}}',
-        '{{first_name}} {{last_name}}', '{{prefix}} {{first_name}} {{last_name}}')
+        '{{first_name}} {{last_name}}',
+        '{{prefix}} {{first_name}} {{last_name}}, {{prefix}} {{first_name}} {{last_name}} {{suffix}}','{{first_name}} {{last_name}} {{suffix}}')
+
+    # Thai prefix
+    # http://www.stou.ac.th/thai/grad_stdy/Apply/prefix.asp
 
     prefixes_female = (
         "พลเอก", "พลโท", "พลตรี", "พันเอก", "พันโท", "พันตรี", "ร้อยเอก", "ร้อยโท",
@@ -34,6 +38,7 @@ class Provider(PersonProvider):
         "ร.ท.", "ร.ต.", "พ.อ.อ.", "พ.อ.ท.", "พ.อ.ต.", "จ.อ.", "จ.ท.", "จ.ต.", "พลฯ",
         "พล.ต.อ.", "พล.ต.ท.", "พล.ต.ต.", "พ.ต.อ.", "พ.ต.ท.", "พ.ต.ต.", "ร.ต.อ.", "ร.ต.ท.",
         "ร.ต.ต.", "ด.ต.", "จ.ส.ต.", "ส.ต.อ.", "ส.ต.ท.", "ส.ต.ต.", "พลฯ", "ม.ล.", "ม.ร.ว.")
+
     prefixes_male = (
         "พลเอก", "พลโท", "พลตรี", "พันเอก", "พันโท", "พันตรี", "ร้อยเอก", "ร้อยโท",
         "ร้อยตรี",
@@ -62,29 +67,64 @@ class Provider(PersonProvider):
 
     prefixes = prefixes_female + prefixes_male
 
-    first_names = (
-        "พรชนก", "พัชชาพลอย", "ภีม", "กชพร", "พัทธพล", "สรัลพร", "จินจุฑา", "สรวุฒิ",
-        "สุวกิจ", "พริมา", "ดาริน", "พินท์ธุสร", "กุลนัน", "ณัฐชา", "ณัฐกฤตา", "สิทธิกร",
-        "วศิน", "อภิสรา", "ณัฐวรรณ", "ธันรดี", "นิรชา", "ธันยชนก", "กมลชนก", "จารุวรรณ",
-        "ผจงรักษ์", "ภัคจิรา", "ปัทธมพร", "สุวิจักษ์", "นพเก้า", "รัชชานนท์", "อธิศ",
-        "คุณาพร", "อริสรา", "ทัศวรรณ", "ชัยวัฒน์", "ภัณฑิรา", "ศุภสุตา", "พัชรพร",
-        "ภูเบศ",
-        "พัทธมน", "ชัญญา", "ปัณณวิชย์", "ชวิน", "ภาดา", "ชนิกานต์", "ณัฐวดี", "ชาลิสา",
[…]
