+++ faker/providers/person/th_TH/__init__.py
@@ -11,11 +11,12 @@ class Provider(PersonProvider):
         '{{first_name}} {{last_name}}',
         '{{first_name}} {{last_name}}',
         '{{first_name}} {{last_name}}',
-        '{{prefix}} {{first_name}} {{last_name}}, {{prefix}} {{first_name}} {{last_name}} {{suffix}}','{{first_name}} {{last_name}} {{suffix}}')
+        '{{prefix}} {{first_name}} {{last_name}}',
+        '{{prefix}} {{first_name}} {{last_name}} {{suffix}}',
+        '{{first_name}} {{last_name}} {{suffix}}')
 
     # Thai prefix
     # http://www.stou.ac.th/thai/grad_stdy/Apply/prefix.asp
-
     prefixes_female = (
         "พลเอก", "พลโท", "พลตรี", "พันเอก", "พันโท", "พันตรี", "ร้อยเอก", "ร้อยโท",
         "ร้อยตรี",
@@ -70,61 +71,84 @@ class Provider(PersonProvider):
     # female first name
     # http: // www.mamaexpert.com / posts / content - 1166
     first_names_female = (
-    "กัสจัง", "เกรซ", "เกรส", "แก้ม", "ข้าวสวย", "ข้าวหอม", "ข้าวทิพย์", "ข้าวปุ้น",
-    "แจน", "ณิชา", "นินิว", "นานา", "นิว", "นับตังค์", "เนเน่", "น้ำมนต์", "โบวี่",
-    "โบนัส", "บิวตี้", "ใบเตย", "ใบบัว", "ใบข้าว", "ใบตอง", "ใบเฟิร์น", "พรีช", "พอใจ",
-    "พอเพียง", "ฟ้าใส", "เฟิร์ส", "เฟสบุค", "เฟสบุ๊ค", "ต้นข้าว", "ต้นหอม", "แตงกว่า",
-    "แตงโม", "ตังค์", "ตังค์", "มิวสิค", "มะนาว", "มิ๊น", "มิ๊นท์", "มินนี่", "บัว",
-    "ปันปัน", "แพท", "แพต", "แพทตี้", "เเพทตี้", "ลูกพีช", "ภูมิใจ", "เอิร์น", "เอิน",
-    "เอวา", "เอมิ", "เอมี่", "ออมสิน", "ออกัส", "ไอซ์",)
+        "กัสจัง", "เกรซ", "เกรส", "แก้ม", "ข้าวสวย", "ข้าวหอม", "ข้าวทิพย์", "ข้าวปุ้น",
+        "แจน", "ณิชา", "นินิว", "นานา", "นิว", "นับตังค์", "เนเน่", "น้ำมนต์", "โบวี่",
+        "โบนัส", "บิวตี้", "ใบเตย", "ใบบัว", "ใบข้าว", "ใบตอง", "ใบเฟิร์น", "พรีช",
+        "พอใจ",
+        "พอเพียง", "ฟ้าใส", "เฟิร์ส", "เฟสบุค", "เฟสบุ๊ค", "ต้นข้าว", "ต้นหอม", "แตงกว่า",
[…]
