### 2017-03-18 f9f59c3
Changed first_name source + add descr
