+++ faker/providers/person/th_TH/__init__.py
@@ -68,47 +68,12 @@ class Provider(PersonProvider):
 
     prefixes = prefixes_female + prefixes_male
 
-    # female first name
-    # http: // www.mamaexpert.com / posts / content - 1166
-    first_names_female = (
-        "กัสจัง", "เกรซ", "เกรส", "แก้ม", "ข้าวสวย", "ข้าวหอม", "ข้าวทิพย์", "ข้าวปุ้น",
-        "แจน", "ณิชา", "นินิว", "นานา", "นิว", "นับตังค์", "เนเน่", "น้ำมนต์", "โบวี่",
-        "โบนัส", "บิวตี้", "ใบเตย", "ใบบัว", "ใบข้าว", "ใบตอง", "ใบเฟิร์น", "พรีช",
-        "พอใจ",
-        "พอเพียง", "ฟ้าใส", "เฟิร์ส", "เฟสบุค", "เฟสบุ๊ค", "ต้นข้าว", "ต้นหอม", "แตงกว่า",
-        "แตงโม", "ตังค์", "ตังค์", "มิวสิค", "มะนาว", "มิ๊น", "มิ๊นท์", "มินนี่", "บัว",
-        "ปันปัน", "แพท", "แพต", "แพทตี้", "เเพทตี้", "ลูกพีช", "ภูมิใจ", "เอิร์น", "เอิน",
-        "เอวา", "เอมิ", "เอมี่", "ออมสิน", "ออกัส", "ไอซ์",)
+    #Chulalongkorn University famous teacher&student first name
+    #https://th.wikipedia.org/wiki/%E0%B8%A3%E0%B8%B2%E0%B8%A2%E0%B8%99%E0%B8%B2%E0%B8%A1%E0%B8%9A%E0%B8%B8%E0%B8%84%E0%B8%84%E0%B8%A5%E0%B8%88%E0%B8%B2%E0%B8%81%E0%B8%88%E0%B8%B8%E0%B8%AC%E0%B8%B2%E0%B8%A5%E0%B8%87%E0%B8%81%E0%B8%A3%E0%B8%93%E0%B9%8C%E0%B8%A1%E0%B8%AB%E0%B8%B2%E0%B8%A7%E0%B8%B4%E0%B8%97%E0%B8%A2%E0%B8%B2%E0%B8%A5%E0%B8%B1%E0%B8%A2
+    first_names = (
[…]
