### 2017-03-18 e30c18f
Fix no_NO building numbers from address provider
Fixes unparsed building number suffixes like 'Drammensveien 1?'
and improves them with an ok distribution of most used suffix letters.
