### 2017-03-18 7229f42
Make it possible to seed uuid4 property #484
