### 2017-04-09 2f5ec43
close #492. Clean up build dir before packaging
