+++ Makefile
@@ -2,6 +2,6 @@ test:
 	coverage run --source=faker setup.py test
 
 release:
-	rm -rf dist
+	rm -rf build dist
 	python setup.py sdist bdist_wheel
 	twine upload dist/*
