### 2017-04-13 0907376
Added ValueError unit test for lorem
