### 2017-04-14 df8e3a2
Add russian words for date_time
