+++ faker/providers/date_time/ru_RU/__init__.py
@@ -1,0 +1,41 @@
+# coding: utf-8
+from __future__ import unicode_literals
+
+from .. import Provider as DateTimeProvider
+
+
+class Provider(DateTimeProvider):
+
+    @classmethod
+    def day_of_week(cls):
+        day = cls.date('%w')
+        DAY_NAMES = {
+            "0": "Воскресенье",
+            "1": "Понедельник",
+            "2": "Вторник",
+            "3": "Среда",
+            "4": "Четверг",
+            "5": "Пятница",
+            "6": "Суббота",
+        }
+        return DAY_NAMES[day]
+
+    @classmethod
+    def month_name(cls):
+        month = cls.month()
+        MONTH_NAMES = {
+            "01": "Январь",
+            "02": "Февраль",
+            "03": "Март",
+            "04": "Апрель",
+            "05": "Май",
+            "06": "Июнь",
+            "07": "Июль",
+            "08": "Август",
+            "09": "Сентябрь",
+            "10": "Октябрь",
+            "11": "Ноябрь",
+            "12": "Декабрь",
+        }
+        return MONTH_NAMES[month]
+
