+++ faker/providers/person/ka_GE/__init__.py
@@ -1,0 +1,153 @@
+# coding=utf-8
+from __future__ import unicode_literals
+
+from .. import Provider as PersonProvider
+
+
+class Provider(PersonProvider):
+    formats_male = (
+        '{{first_name_male}} {{last_name}}',
+    )
+    formats_female = (
+        '{{first_name_female}} {{last_name}}',
+    )
+
+    formats = formats_male + formats_female
+
+    first_names_male = (
+        'ავთანდილ', 'აკაკი', 'ალექს', 'ალექსანდრე', 'ამირან', 'ანზორ', 'არმენ',
+        'არტურ', 'არჩილ', 'ასლან', 'ბადრი', 'ბესარიონ', 'ბესიკ', 'ბექა',
+        'ბიძინა', 'ბონდო', 'ბორის', 'გელა', 'გენადი', 'გია', 'გიგა', 'გიგლა',
+        'გივი', 'გიორგი', 'გოგა', 'გოგი', 'გოგიტა', 'გოდერძი', 'გოჩა', 'გრიგოლ',
+        'გურამ', 'დავით', 'დათო', 'დემურ', 'დიმიტრი', 'ედუარდ', 'ელგუჯა',
+        'ემზარ', 'ვალერი', 'ვალერიან', 'ვანო', 'ვაჟა', 'ვასილ', 'ვახტანგ',
+        'ვახტანგ', 'ვეფხვია', 'ვიქტორ', 'ვლადიმერ', 'ზაზა', 'ზაურ', 'ზაქარია',
+        'ზვიად', 'ზურაბ', 'თამაზ', 'თეიმურაზ', 'თემურ', 'თენგიზ', 'თორნიკე',
+        'იაგო', 'ივანე', 'ილია', 'იოსებ', 'ირაკლი', 'იური', 'კარლო', 'კახა',
+        'კახაბერ', 'კობა', 'კონსტანტინე', 'ლაშა', 'ლევან', 'მალხაზ', 'მამუკა',
+        'მერაბ', 'მინდია', 'მირიან', 'მიხეილ', 'მიხეილ', 'მურთაზ', 'მურმან',
+        'ნიკა', 'ნიკოლოზ', 'ნოდარ', 'ნუგზარ', 'ნუკრი', 'ოთარ', 'ომარ', 'პაატა',
+        'პავლე', 'პეტრე', 'რამაზ', 'რატი', 'რევაზ', 'რეზო', 'რობერტ', 'როინ',
+        'როლანდ', 'რომან', 'სერგო', 'სიმონ', 'სოსო', 'ტარიელ', 'უშანგი', 'უჩა',
+        'შალვა', 'შაქრო', 'შოთა', 'ხვიჩა', 'ჯაბა', 'ჯემალ', 'ჯონი', 'ჯუმბერ'
+    )
+
+    first_names_female = (
+        'აზა', 'აიდა', 'ალა', 'ანა', 'ანი', 'ანიკო', 'ანნა', 'ანჟელა', 'ასმათ',
+        'ბელა', 'გალინა', 'გვანცა', 'გიული', 'გუგული', 'გულიკო', 'გულნარა',
+        'დალი', 'დარეჯან', 'დიანა', 'დოდო', 'ევგენია', 'ეთერ', 'ეკა',
+        'ეკატერინე', 'ელენა', 'ელენე', 'ელზა', 'ელიზა', 'ელისო', 'ელმირა',
[…]
