### 2017-04-20 5f5fb47
Added unit test for new_date method
