### 2017-04-22 df4dc60
Added company provider to hu_HU locale
