### 2017-04-24 b656a55
Fix typo
Replace unicode '×' symbol by x litter in image placeholder url
