### 2017-04-29 1ddd15e
Improved hu_HU accuracy
