### 2017-05-09 083bedd
Implemented hu_HU months + days
