### 2017-05-09 #94 @sckott
no, not waiting on answers. have addressed many issues so far, but a few more to go (label: review)  https://github.com/ropenscilabs/charlatan/issues?q=is%3Aissue+label%3Areview+is%3Aopen
