### 2017-05-10 26ed12d
Replacement rules for emails à->a, è->e
