+++ faker/providers/internet/de_DE/__init__.py
@@ -15,5 +15,7 @@ class Provider(InternetProvider):
         ('ä', 'ae'), ('Ä', 'Ae'),
         ('ö', 'oe'), ('Ö', 'Oe'),
         ('ü', 'ue'), ('Ü', 'Ue'),
+        ('é', 'e'), ('É', 'E'),
+        ('à', 'a'), ('À', 'A'),
         ('ß', 'ss'),
     )
