### 2017-05-10 f3a9d84
Bump version: 0.7.11 → 0.7.12
