+++ faker/__init__.py
@@ -1,4 +1,4 @@
-VERSION = '0.7.11'
+VERSION = '0.7.12'
 
 from faker.generator import Generator
 from faker.factory import Factory
