### 2017-05-16 91a2ca8
Removed capitalisation from hu_HU address
