+++ faker/providers/person/et_EE/__init__.py
@@ -1,0 +1,190 @@
+# coding=utf-8
+from __future__ import unicode_literals
+from .. import Provider as PersonProvider
+
+
+class Provider(PersonProvider):
+    # https://en.wikipedia.org/wiki/Demographics_of_Estonia#Ethnic_groups
+    # Main population groups in Estonia are Estonians and ethnic Russians:
+    # About 70% of the population are Estonians and about 25% are Russians
+    est_rat = 0.7
+    rus_rat = 1.0 - est_rat
+    formats = {'{{first_name_est}} {{last_name_est}}': est_rat,
+               '{{first_name_rus}} {{last_name_rus}}': rus_rat}
+
+    formats_male = {'{{first_name_male_est}} {{last_name_est}}': est_rat,
+                    '{{first_name_male_rus}} {{last_name_rus}}': rus_rat}
+    formats_female = {'{{first_name_female_est}} {{last_name_est}}': est_rat,
+                      '{{first_name_female_rus}} {{last_name_rus}}': rus_rat}
+
+    prefixes_neutral = ('doktor', 'dr', 'prof')
+    prefixes_male = ('härra', 'hr') + prefixes_neutral
+    prefixes_female = ('proua', 'pr') + prefixes_neutral
+    prefixes = set(prefixes_male + prefixes_female)
+
+    suffixes = ('PhD', 'MSc', 'BSc')
+
+    # source: http://www.stat.ee/public/apps/nimed/TOP
+    # TOP 50 male names in 2017 according to the Statistics Estonia
+    first_names_male_est = ('Aivar', 'Aleksander', 'Alexander', 'Andres',
+                            'Andrus', 'Ants', 'Indrek', 'Jaan', 'Jaanus',
+                            'Jüri', 'Kristjan', 'Marek', 'Margus', 'Marko',
+                            'Martin', 'Mati', 'Meelis', 'Mihkel',  'Peeter',
+                            'Priit', 'Raivo', 'Rein', 'Sander', 'Siim', 'Tarmo',
+                            'Tiit', 'Toomas', 'Tõnu', 'Urmas', 'Vello')
+
+    first_names_female_est = ('Aino', 'Anna', 'Anne', 'Anneli', 'Anu', 'Diana',
+                              'Ene', 'Eve', 'Kadri', 'Katrin', 'Kristi',
+                              'Kristiina', 'Kristina', 'Laura', 'Linda', 'Maie',
+                              'Malle', 'Mare', 'Maria', 'Marika', 'Merike',
+                              'Niina', 'Piret', 'Reet', 'Riina', 'Sirje',
+                              'Tiina', 'Tiiu', 'Triin', 'Ülle')
+
+    first_names_est = first_names_male_est + first_names_female_est
+
+    first_names_male_rus = ('Aleksander', 'Aleksandr', 'Aleksei', 'Alexander',
+                            'Andrei', 'Artur', 'Dmitri', 'Igor', 'Ivan',
+                            'Jevgeni', 'Juri', 'Maksim', 'Mihhail', 'Nikolai',
+                            'Oleg', 'Pavel', 'Roman', 'Sergei', 'Sergey',
+                            'Valeri', 'Viktor', 'Vladimir')
+
+    first_names_female_rus = ('Aleksandra', 'Anna', 'Diana', 'Elena', 'Galina',
+                              'Irina', 'Jekaterina', 'Jelena', 'Julia',
[…]
