### 2017-05-29 e2eb12b
Fix diacritics in pl_PL street names
