+++ faker/providers/address/pl_PL/__init__.py
@@ -54,86 +54,86 @@ class Provider(AddressProvider):
     )
 
     streets = (
-        'Polna', 'Lesna', 'Sloneczna', 'Krótka', 'Szkolna', 'Ogrodowa',
-        'Lipowa', 'Brzozowa', 'Lakowa', 'Kwiatowa', 'Sosnowa', 'Koscielna',
-        'Akacjowa', 'Parkowa', 'Zielona', 'Kolejowa', 'Sportowa', 'Debowa',
-        'Kosciuszki', 'Maja', 'Mickiewicza', 'Cicha', 'Spokojna', 'Klonowa',
+        'Polna', 'Leśna', 'Słoneczna', 'Krótka', 'Szkolna', 'Ogrodowa',
+        'Lipowa', 'Brzozowa', 'Łąkowa', 'Kwiatowa', 'Sosnowa', 'Kościelna',
+        'Akacjowa', 'Parkowa', 'Zielona', 'Kolejowa', 'Sportowa', 'Dębowa',
+        'Kościuszki', 'Maja', 'Mickiewicza', 'Cicha', 'Spokojna', 'Klonowa',
         'Spacerowa', 'Swierkowa', 'Kasztanowa', 'Nowa', 'Piaskowa',
-        'Sienkiewicza', 'Rózana', 'Topolowa', 'Wisniowa', 'Dworcowa',
-        'Wiejska', 'Graniczna', 'Slowackiego', 'Dluga', 'Wrzosowa',
-        'Konopnickiej', 'Boczna', 'Waska', 'Wierzbowa', 'Jasminowa',
-        'Wspólna', 'Modrzewiowa', 'Kopernika', 'Jana Pawla II',
-        'Poprzeczna', 'Wesola', 'Pogodna', 'Zeromskiego', 'Rynek', 'Bukowa',
-        'Wojska Polskiego', 'Sadowa', 'Górna', 'Jodlowa', 'Wolnosci',
-        'Glówna', 'Mlynska', 'Strazacka', 'Prusa', 'Jesionowa', 'Przemyslowa',
-        'Osiedlowa', 'Wiosenna', 'Sikorskiego', 'Chopina', 'Poludniowa',
-        'Malinowa', 'Stawowa', 'Reymonta', 'Pilsudskiego', 'Zacisze',
-        'Cmentarna', 'Okrezna', 'Kochanowskiego', 'Armii Krajowej', 'Mila',
+        'Sienkiewicza', 'Rózana', 'Topolowa', 'Wiśniowa', 'Dworcowa',
+        'Wiejska', 'Graniczna', 'Słowackiego', 'Długa', 'Wrzosowa',
+        'Konopnickiej', 'Boczna', 'Wąska', 'Wierzbowa', 'Jaśminowa',
+        'Wspólna', 'Modrzewiowa', 'Kopernika', 'Jana Pawła II',
+        'Poprzeczna', 'Wesoła', 'Pogodna', 'Żeromskiego', 'Rynek', 'Bukowa',
+        'Wojska Polskiego', 'Sadowa', 'Górna', 'Jodłowa', 'Wolności',
+        'Glówna', 'Młyńska', 'Strażacka', 'Prusa', 'Jesionowa', 'Przemysłowa',
+        'Osiedlowa', 'Wiosenna', 'Sikorskiego', 'Chopina', 'Południowa',
+        'Malinowa', 'Stawowa', 'Reymonta', 'Piłsudskiego', 'Zacisze',
+        'Cmentarna', 'Okrężna', 'Kochanowskiego', 'Armii Krajowej', 'Miła',
         'Jasna', 'Wodna', 'Zamkowa', 'Witosa', 'Reja', 'Warszawska',
[…]
