### 2017-05-29 1f2feb1
Missing test for pl_PL dow and month name
