### 2017-05-30 18d6b8a
Make static data accessible at class level
This cleans tests by giving direct access to dataset used to generate results
