+++ tests/providers/date_time.py
@@ -24,38 +24,14 @@ class TestHuHU(unittest.TestCase):
 
 class TestPlPL(unittest.TestCase):
 
-    DAY_NAMES = (
-        'poniedziałek',
-        'wtorek',
-        'środa',
-        'czwartek',
-        'piątek',
-        'sobota',
-        'niedziela',
-    )
-
-    MONTH_NAMES = (
-        'styczeń',
-        'luty',
-        'marzec',
-        'kwiecień',
-        'maj',
-        'czerwiec',
-        'lipiec',
-        'sierpień',
-        'wrzesień',
-        'październik',
-        'listopad',
-        'grudzień'
-    )
-
     def setUp(self):
         self.factory = Factory.create('pl_PL')
+        self.provider = self.factory.provider('faker.providers.date_time')
 
     def test_day(self):
         day = self.factory.day_of_week()
-        assert day in self.DAY_NAMES
+        assert day in self.provider.DAY_NAMES.values()
 
     def test_month(self):
         month = self.factory.month_name()
-        assert month in self.MONTH_NAMES
+        assert month in self.provider.MONTH_NAMES.values()
