### 2017-05-31 #94 @sckott
@tjmahr @geanders Okay, sorry for the long delay.

I've now dealt with nearly all of your points - see issues with `review` label https://github.com/ropenscilabs/charlatan/milestone/1?closed=1  - associated with the `v0.1` milestone - for my first cran push of the pkg

I've moved a few of your issues/or parts of your issues to the next milestone  https://github.com/ropenscilabs/charlatan/milestone/2 - let me know if you think they're important enough we need to deal with them before going to CRAN.

and @noamross let me know if you see anything I should fix/make better
