### 2017-05-31 eac547e
Fix for issue #513 in test test_date_time_between_dates_with_tzinfo
Problem was in function datetime_to_timestamp from faker/providers/date_time/__init__.py
time.mktime() assumes that the passed tuple is in local time and calendar.timegm() assumes it's in UTC.
 After fix in this commit there are no such issue.
