### 2017-06-02 f78aaf7
Fix for issue #513 in test test_date_time_between_dates_with_tzinfo
Change converting datetime to timestamp from mktime with local timezone to calendar.timegm with UTC.
It fix bug for some countries which used decret time (timezone +1 hour).
