### 2017-06-02 e155a10
Add `future_date`, `future_datetime`, `past_date` and `past_datetime`
