+++ faker/__init__.py
@@ -1,4 +1,4 @@
-VERSION = '0.7.12'
+VERSION = '0.7.13'
 
 from faker.generator import Generator
 from faker.factory import Factory
