### 2017-06-02 3df9e0b
fix `past_date` and `past_datetime` methods
