### 2017-06-02 ef9a3ab
fix docstring
