+++ faker/__init__.py
@@ -1,4 +1,4 @@
-VERSION = '0.7.13'
+VERSION = '0.7.14'
 
 from faker.generator import Generator
 from faker.factory import Factory
