### 2017-06-02 c623dee
fix start and end date for datetime provider methods
