+++ faker/__init__.py
@@ -1,4 +1,4 @@
-VERSION = '0.7.14'
+VERSION = '0.7.15'
 
 from faker.generator import Generator
 from faker.factory import Factory
