### 2017-06-03 fbfdb61
add test
need to try with py.test
