### 2017-06-03 c895c6a
documentation and tests
