### 2017-06-03 991a84e
try to render correctly the readme.rst
