### 2017-06-03 1c497ae
adding content to the readme.rst
