### 2017-06-04 #94 @noamross
I am OK with the items moved to the 0.2 milestone if the reviewers are.

I note that test coverage is still modest, including in some areas (colors), where your reviewers identified bugs or inconsistencies. Can you bring that up?

@tjmahr and @geanders, please take a look at the updated version and let us know if @sckott has addressed your comments to your satisfaction.
