### 2017-06-05 #94 @geanders
This has satisfactorily addressed my comments, and I'm ok with the items mover to the 0.2 milestones, as well.
