### 2017-06-09 d89c3bd
simplifying other functions
