### 2017-06-09 bdc9806
add the return element in the doctring
