### 2017-06-09 2a0d2a5
simplify fake.sentences()
