### 2017-06-10 #94 @sckott
thanks for the PR @tjmahr

ready to make quick fixes indeed!
