### 2017-06-11 #94 @noamross
Thank you for your responses, @geanders and @tjmahr!

@sckott: This looks good as soon as you finish addressing https://github.com/ropenscilabs/charlatan/pull/47 .
