+++ tests/factory.py
@@ -2,9 +2,7 @@
 
 from __future__ import unicode_literals
 
-import datetime
 import re
-import time
 import unittest
 import string
 import six
@@ -21,32 +19,14 @@ from faker.utils import text, decorators
 from . import string_types
 
 
-class UTC(datetime.tzinfo):
-    """
-    UTC implementation taken from Python's docs.
-    """
-    def __repr__(self):
-        return "<UTC>"
-
-    def utcoffset(self, dt):
-        return datetime.timedelta(0)
-
-    def tzname(self, dt):
-        return "UTC"
-
-    def dst(self, dt):
-        return datetime.timedelta(0)
-
-
-utc = UTC()
-
-
 class BarProvider(object):
+
     def foo_formatter(self):
         return 'barfoo'
 
 
 class FooProvider(object):
+
     def foo_formatter(self):
         return 'foobar'
 
@@ -181,179 +161,13 @@ class FactoryTestCase(unittest.TestCase):
         pick = provider.random_element(choices)
         self.assertTrue(pick in choices)
 
-        choices = {'a': 5, 'b': 2, 'c': 2, 'd':1 }
+        choices = {'a': 5, 'b': 2, 'c': 2, 'd': 1}
         pick = provider.random_element(choices)
         self.assertTrue(pick in choices)
 
-
-        choices = {'a': 0.5, 'b': 0.2, 'c': 0.2, 'd':0.1}
+        choices = {'a': 0.5, 'b': 0.2, 'c': 0.2, 'd': 0.1}
         pick = provider.random_element(choices)
         self.assertTrue(pick in choices)
-
-    def test_timezone_conversion(self):
-        from faker.providers.date_time import datetime_to_timestamp
-
-        now = datetime.datetime.now(utc).replace(microsecond=0)
-        timestamp = datetime_to_timestamp(now)
-        now_back = datetime.datetime.fromtimestamp(timestamp, utc)
-        self.assertEqual(now, now_back)
-
-        today = datetime.date.today()
-        timestamp = datetime_to_timestamp(today)
-        today_as_dt = datetime.datetime.combine(today, datetime.time.min)
-        today_back = datetime.datetime.fromtimestamp(timestamp)
-        self.assertEqual(today_as_dt, today_back)
-
-    def test_datetime_safe(self):
-        from faker.utils import datetime_safe
-        # test using example provided in module
-        result = datetime_safe.date(1850, 8, 2).strftime('%Y/%m/%d was a %A')
-        self.assertEqual(result, '1850/08/02 was a Friday')
-        # test against certain formatting strings used on pre-1900 dates
-        with self.assertRaises(TypeError):
-            datetime_safe.date(1850, 8, 2).strftime('%s')
-        with self.assertRaises(TypeError):
-            datetime_safe.date(1850, 8, 2).strftime('%y')
-        # test using 29-Feb-2012 and escaped percentage sign
-        result = datetime_safe.date(2012, 2, 29).strftime('%Y-%m-%d was a 100%% %A')
-        self.assertEqual(result, r'2012-02-29 was a 100% Wednesday')
-        # test that certain formatting strings are allowed on post-1900 dates
-        result = datetime_safe.date(2008, 2, 29).strftime('%y')
-        self.assertEqual(result, r'08')
-
-    def test_datetime_safe_new_date(self):
-        from faker.utils import datetime_safe
-        d = datetime_safe.date(1850, 8, 2)
-        result = datetime_safe.new_date(d)
-        self.assertEqual(result, datetime.date(1850, 8, 2))
-
-    def test_datetimes_with_and_without_tzinfo(self):
-        from faker.providers.date_time import Provider
-        provider = Provider
-
[…]
