+++ tests/providers/date_time.py
@@ -1,12 +1,36 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
-from datetime import date, datetime
+from datetime import date, datetime, timedelta, tzinfo
+from datetime import time as datetime_time
+import time
 import unittest
 
 from faker import Factory
+from faker.generator import random
 from faker.providers.date_time import Provider as DatetimeProvider
 from .. import string_types
+
+
+class UTC(tzinfo):
+    """
+    UTC implementation taken from Python's docs.
+    """
+
+    def __repr__(self):
+        return "<UTC>"
+
+    def utcoffset(self, dt):
+        return timedelta(0)
+
+    def tzname(self, dt):
+        return "UTC"
+
+    def dst(self, dt):
+        return timedelta(0)
+
+
+utc = UTC()
 
 
 class TestDateTime(unittest.TestCase):
@@ -42,6 +66,171 @@ class TestDateTime(unittest.TestCase):
         timestamp = DatetimeProvider._parse_date_time('+30d')
         now = DatetimeProvider._parse_date_time('now')
         self.assertTrue(timestamp > now)
+
+    def test_timezone_conversion(self):
+        from faker.providers.date_time import datetime_to_timestamp
+
+        now = datetime.now(utc).replace(microsecond=0)
+        timestamp = datetime_to_timestamp(now)
+        now_back = datetime.fromtimestamp(timestamp, utc)
+        self.assertEqual(now, now_back)
+
+        today = date.today()
+        timestamp = datetime_to_timestamp(today)
+        today_as_dt = datetime.combine(today, datetime_time.min)
+        today_back = datetime.fromtimestamp(timestamp)
+        self.assertEqual(today_as_dt, today_back)
+
+    def test_datetime_safe(self):
+        from faker.utils import datetime_safe
+        # test using example provided in module
+        result = datetime_safe.date(1850, 8, 2).strftime('%Y/%m/%d was a %A')
+        self.assertEqual(result, '1850/08/02 was a Friday')
+        # test against certain formatting strings used on pre-1900 dates
+        with self.assertRaises(TypeError):
+            datetime_safe.date(1850, 8, 2).strftime('%s')
+        with self.assertRaises(TypeError):
+            datetime_safe.date(1850, 8, 2).strftime('%y')
+        # test using 29-Feb-2012 and escaped percentage sign
+        result = datetime_safe.date(2012, 2, 29).strftime('%Y-%m-%d was a 100%% %A')
+        self.assertEqual(result, r'2012-02-29 was a 100% Wednesday')
+        # test that certain formatting strings are allowed on post-1900 dates
+        result = datetime_safe.date(2008, 2, 29).strftime('%y')
+        self.assertEqual(result, r'08')
+
+    def test_datetime_safe_new_date(self):
+        from faker.utils import datetime_safe
+        d = datetime_safe.date(1850, 8, 2)
+        result = datetime_safe.new_date(d)
+        self.assertEqual(result, date(1850, 8, 2))
+
+    def test_datetimes_with_and_without_tzinfo(self):
+        from faker.providers.date_time import Provider
+        provider = Provider
+
+        self.assertEqual(provider.date_time().tzinfo, None)
+        self.assertEqual(provider.date_time(utc).tzinfo, utc)
+
+        self.assertEqual(provider.date_time_ad().tzinfo, None)
+        self.assertEqual(provider.date_time_ad(utc).tzinfo, utc)
+
+        self.assertFalse(provider.iso8601().endswith('+00:00'))
+        self.assertTrue(provider.iso8601(utc).endswith('+00:00'))
+
+    def test_date_object(self):
+        from faker.providers.date_time import Provider
+        provider = Provider
+
+        self.assertIsInstance(provider.date_object(), date)
+
+    def test_time_object(self):
[…]
