### 2017-06-12 92622bf
fix test for timezone conversion
