+++ docs/conf.py
@@ -54,9 +54,9 @@ copyright = u'2014, Daniele Faraglia'
 # built documents.
 #
 # The short X.Y version.
-version = '0.7.16'
+version = '0.7.17'
 # The full version, including alpha/beta/rc tags.
-release = '0.7.16'
+release = '0.7.17'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
