+++ faker/__init__.py
@@ -1,4 +1,4 @@
-VERSION = '0.7.16'
+VERSION = '0.7.17'
 
 from faker.generator import Generator
 from faker.factory import Factory
