### 2017-06-15 #94 @noamross
Approved! @geanders and @geanders for your thorough work throughout this review.  Good work @sckott, feel free to transfer to the ropensci repository, and consider making a post for the blog.
