### 2017-06-15 #94 @sckott
thanks everyone for your work that helped to improve the package!
