### 2017-06-15 c846dee
Update __init__.py
Using timestamp doesn't allow you to generate a date before 1970. In fact, you can even put in (-100y,-90y) as parameters and get a runtime error.
