### 2017-06-16 5cdf229
Update __init__.py
added timezone back in
%Z returning correct timezone back
print(faker.date_time_between('-20y','now',tzinfo).strftime('%Y-%m-%d %H:%M:%S %Z'))
