### 2017-06-16 e0cb001
Update __init__.py
Fixes #488 and #530
