### 2017-06-16 0072cd9
Update __init__.py
bytes type is python 3, whereas bytesarray is 2. Need to add check for version 2 to use bytearray instead, otherwise bytes in python 2 is simply an alias to str
