### 2017-06-16 1648f5d
Add in seed test for binary and fixed isinstance
Pull #535
