### 2017-06-16 46840a6
Update factory.py
Fixed a typo
