### 2017-06-16 dab3c3d
Fixed test_binary
