### 2017-06-16 5aa31c9
fix typo in tests
