### 2017-06-17 17ebd06
Added color pt_BR localization.
