+++ faker/providers/color/pt_BR/__init__.py
@@ -1,0 +1,260 @@
+# coding=utf-8
+
+from __future__ import unicode_literals
+from collections import OrderedDict
+
+from .. import Provider as ColorProvider
+
+localized = True
+
+
+class Provider(ColorProvider):
+    all_colors = OrderedDict((
+        ("Água", "#00FFFF"),
+        ("Azul turquesa", "#00CCEE"),
+        ("Azul turquesa brilhante", "#00DDFF"),
+        ("Azul brasilis", "#00BDCE"),
+        ("Azul brasilis brilhante", "#09ACDB"),
+        ("Água-marinha", "#7FFFD4"),
+        ("Água-marinha média", "#66CDAA"),
+        ("Azul celeste pernambucano", "#00A4CD"),
+        ("Alizarina", "#E32636"),
+        ("Amarelo", "#FFFF00"),
+        ("Amarelo brasilis", "#ECDB00"),
+        ("Amarelo creme", "#ECD690"),
+        ("Amarelo claro", "#FFFFE0"),
+        ("Amarelo escuro", "#F2B73F"),
+        ("Amarelo esverdeado", "#ADFF2F"),
+        ("Amarelo ouro claro", "#FAFAD2"),
+        ("Amarelo queimado", "#EEAD2D"),
+        ("Âmbar", "#FFBF00"),
+        ("Ameixa", "#DDA0DD"),
+        ("Ametista", "#9966CC"),
+        ("Amêndoa", "#FFEBCD"),
+        ("Aspargo", "#7BA05B"),
+        ("Azul", "#0000FF"),
+        ("Azul alice", "#F0F8FF"),
+        ("Azul ardósia", "#6A5ACD"),
+        ("Azul ardósia claro", "#8470FF"),
+        ("Azul ardósia escuro", "#483D8B"),
+        ("Azul ardósia médio", "#7B68EE"),
+        ("Azul areado", "#B8CAD4"),
+        ("Azul aço", "#4682B4"),
+        ("Azul aço claro", "#B0C4DE"),
+        ("Azul cadete", "#5F9EA0"),
+        ("Azul camarada", "#054F77"),
+        ("Azul celeste", "#F0FFFF"),
+        ("Azul celeste brilhante", "#007FFF"),
+        ("Azul claro", "#ADD8E6"),
+        ("Azul cobalto", "#0047AB"),
+        ("Azul céu", "#87CEEB"),
+        ("Azul céu claro", "#87CEFA"),
+        ("Azul céu profundo", "#00BFFF"),
+        ("Azul escuro", "#00008B"),
+        ("Azul flor de milho", "#6495ED"),
+        ("Azul força aérea", "#5D8AA8"),
+        ("Azul furtivo", "#1E90FF"),
+        ("Azul manteiga", "#a6aa3e"),
+        ("Azul marinho", "#120A8F"),
+        ("Azul meia-noite", "#191970"),
+        ("Azul médio", "#0000CD"),
+        ("Azul petróleo", "#084D6E"),
+        ("Azul pólvora", "#B0E0E6"),
+        ("Azul taparuere", "#248EFF"),
+        ("Azul real", "#0000DD"),
+        ("Azul violeta", "#8A2BE2"),
+        ("Açafrão", "#F4C430"),
+        ("Bege", "#F5F5DC"),
+        ("Bordô", "#800000"),
+        ("Borgonha", "#900020"),
+        ("Branco", "#FFFFFF"),
+        ("Branco antigo", "#FAEBD7"),
+        ("Branco fantasma", "#F8F8FF"),
+        ("Branco floral", "#FFFAF0"),
[…]
