### 2017-06-18 a43e1c0
Adding an implementation of `ssn` provider for the pl_PL locale.
