+++ tests/providers/ssn.py
@@ -4,11 +4,14 @@ from __future__ import unicode_literals
 
 import unittest
 import re
+from datetime import datetime
 
 from faker import Factory
 from faker.providers.ssn.et_EE import Provider as EtProvider, checksum as et_checksum
 from faker.providers.ssn.hr_HR import Provider as HrProvider, checksum as hr_checksum
 from faker.providers.ssn.pt_BR import Provider as PtProvider, checksum as pt_checksum
+from faker.providers.ssn.pl_PL import (Provider as PlProvider, checksum as pl_checksum,
+                                       calculate_month as pl_calculate_mouth)
 
 
 class TestEtEE(unittest.TestCase):
@@ -48,7 +51,6 @@ class TestHrHR(unittest.TestCase):
 
 
 class TestPtBR(unittest.TestCase):
-
     def setUp(self):
         self.factory = Factory.create('pt_BR')
 
@@ -64,3 +66,41 @@ class TestPtBR(unittest.TestCase):
         for _ in range(100):
             self.assertTrue(re.search(r'\d{3}\.\d{3}\.\d{3}\-\d{2}', PtProvider.cpf()))
 
+
+class TestPlPL(unittest.TestCase):
+    """ Tests SSN in the pl_PL locale """
+
+    def setUp(self):
+        self.factory = Factory.create('pl_PL')
+
+    def test_ssn_checksum(self):
+        self.assertEqual(pl_checksum([0, 5, 2, 6, 2, 8, 1, 2, 3, 6]), 5)
+        self.assertEqual(pl_checksum([8, 5, 0, 5, 0, 8, 1, 5, 5, 8]), 7)
+        self.assertEqual(pl_checksum([4, 5, 1, 1, 1, 0, 0, 2, 4, 3]), 3)
+        self.assertEqual(pl_checksum([9, 1, 0, 7, 2, 6, 1, 4, 8, 7]), 3)
+        self.assertEqual(pl_checksum([8, 1, 1, 2, 1, 4, 1, 1, 8, 7]), 6)
+
+    def test_calculate_month(self):
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 1800', '%m %d %Y')), 81)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('12 1 1800', '%m %d %Y')), 92)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 1899', '%m %d %Y')), 81)
+
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 1900', '%m %d %Y')), 1)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('12 1 1900', '%m %d %Y')), 12)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 1999', '%m %d %Y')), 1)
+
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 2000', '%m %d %Y')), 21)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('12 1 2000', '%m %d %Y')), 32)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 2099', '%m %d %Y')), 21)
+
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 2100', '%m %d %Y')), 41)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('12 1 2100', '%m %d %Y')), 52)
+        self.assertEqual(pl_calculate_mouth(datetime.strptime('1 1 2199', '%m %d %Y')), 41)
+
[…]
