### 2017-06-18 7acd48c
Removal of  year 1800 test case
