### 2017-06-18 a579db0
Added a pt_BR localization for colors provider.
