+++ faker/providers/color/pt_BR/__init__.py
@@ -10,108 +10,108 @@ localized = True
 
 class Provider(ColorProvider):
     all_colors = OrderedDict((
-        ("Água", "#00FFFF"),
-        ("Azul turquesa", "#00CCEE"),
-        ("Azul turquesa brilhante", "#00DDFF"),
-        ("Azul brasilis", "#00BDCE"),
-        ("Azul brasilis brilhante", "#09ACDB"),
-        ("Água-marinha", "#7FFFD4"),
+        ("Açafrão", "#F4C430"),
         ("Água-marinha média", "#66CDAA"),
-        ("Azul celeste pernambucano", "#00A4CD"),
+        ("Água-marinha", "#7FFFD4"),
+        ("Água", "#00FFFF"),
         ("Alizarina", "#E32636"),
-        ("Amarelo", "#FFFF00"),
         ("Amarelo brasilis", "#ECDB00"),
-        ("Amarelo creme", "#ECD690"),
         ("Amarelo claro", "#FFFFE0"),
+        ("Amarelo creme", "#ECD690"),
         ("Amarelo escuro", "#F2B73F"),
+        ("Amarelo esverdeado", "#9ACD32"),
         ("Amarelo esverdeado", "#ADFF2F"),
         ("Amarelo ouro claro", "#FAFAD2"),
         ("Amarelo queimado", "#EEAD2D"),
+        ("Amarelo", "#FFFF00"),
         ("Âmbar", "#FFBF00"),
         ("Ameixa", "#DDA0DD"),
-        ("Ametista", "#9966CC"),
         ("Amêndoa", "#FFEBCD"),
+        ("Ametista", "#9966CC"),
         ("Aspargo", "#7BA05B"),
-        ("Azul", "#0000FF"),
+        ("Azul aço claro", "#B0C4DE"),
+        ("Azul aço", "#4682B4"),
         ("Azul alice", "#F0F8FF"),
-        ("Azul ardósia", "#6A5ACD"),
         ("Azul ardósia claro", "#8470FF"),
         ("Azul ardósia escuro", "#483D8B"),
         ("Azul ardósia médio", "#7B68EE"),
+        ("Azul ardósia", "#6A5ACD"),
         ("Azul areado", "#B8CAD4"),
-        ("Azul aço", "#4682B4"),
-        ("Azul aço claro", "#B0C4DE"),
+        ("Azul brasilis brilhante", "#09ACDB"),
+        ("Azul brasilis", "#00BDCE"),
         ("Azul cadete", "#5F9EA0"),
         ("Azul camarada", "#054F77"),
-        ("Azul celeste", "#F0FFFF"),
         ("Azul celeste brilhante", "#007FFF"),
-        ("Azul claro", "#ADD8E6"),
-        ("Azul cobalto", "#0047AB"),
-        ("Azul céu", "#87CEEB"),
+        ("Azul celeste pernambucano", "#00A4CD"),
+        ("Azul celeste", "#F0FFFF"),
         ("Azul céu claro", "#87CEFA"),
         ("Azul céu profundo", "#00BFFF"),
+        ("Azul céu", "#87CEEB"),
+        ("Azul claro", "#ADD8E6"),
+        ("Azul cobalto", "#0047AB"),
         ("Azul escuro", "#00008B"),
         ("Azul flor de milho", "#6495ED"),
         ("Azul força aérea", "#5D8AA8"),
         ("Azul furtivo", "#1E90FF"),
         ("Azul manteiga", "#a6aa3e"),
         ("Azul marinho", "#120A8F"),
-        ("Azul meia-noite", "#191970"),
         ("Azul médio", "#0000CD"),
+        ("Azul meia-noite", "#191970"),
[…]
