### 2017-06-19 99a4abf
Create a method for codes of cryptocurrencies in the currency provider
